Guesses the domain name of the ISP of a host
In addition to testing Internet bandwidth, which I'll probably do using speedtest-cli (python), I want to test LAN throughput, which may be doable if the scripts figure out the ISP's website and clocks the download speed of several of its largest files. Hopefully I'll figure out a better way than this to get the domain.